requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pandas==2.2.3
scikit-learn==1.5.2
lightgbm==4.5.0
xgboost==2.1.1
python-multipart==0.0.10
pydantic==2.9.2
pydantic-settings==2.5.2
PyJWT==2.8.0

fastapi-users[sqlalchemy]
passlib[bcrypt] 
python-jose[cryptography]